Building MVP classes
Implemented code for ViewModel, Activity, MVP interface, Presenter, and ListAdapter.
Adjusted code from the loadData method from the Presenter to use the RxJava2 library as the course example used RxJava1

In RxJava2 there is no Subscription class. We now use Disposable class in cojunction with subscribeWith and use DisposableObserver instead of Observer (because we are using Observable model object). We also use dispose() method to close "subscription" instead of unsubscribe().